requirements.txt
opencv-python
mediapipe
pyautogui
numpy
Pillow
pygame
